Fixed port bug
Fixed a bug where opening a port opens it for ALL nodes.
This was because data.getPort() returned the port instead of a copy.
The programs.PortBreakingTool would open the port, and as a result, this would reflect everywhere.
Returning a copy makes sure each node has a separate port.