Added save file system
The game can now be saved using the SAVE command. It uses Python's CONFIGFILE module under the hood. It even automatically generates file names. There is a chance that a save file can be over-written, but the chance of that is 1 in 2^64. Perhaps I'll devise a better system in future.