Added time into save game name generator
The old algorithm worked like this:

1. Generate a number between 0 and 2^64 - 1
2. Encode it with SHA256

This is fine, but introduces the (small) risk of the same number being
picked for a second save file, leading to an older one being
over-written.

The current algorithm works like this:

1. Generate a number between 0 and 2^64 - 1
2. Multiply that by the current time (courtesy of time.time())
3. Encode it with SHA256

This is much more secure and basically eliminates the possibility of a
collision happening, because time.time() has accuracy down to the
microseconds on some operating systems. For a collision to happen,
you'd need to set your clock backwards to the EXACT time you saved, and
get it accurate to the MICROSECOND, and you'd have a
0.000000000000000005% chance of a collision. The time basically acts as
  a seed of sorts, courtesy of the operating system's kernel.

Q: Is this overkill?
A: Probably.

Q: Couldn't you just check what files are in the save game directory?
A: That wouldn't work for updating a save file.

Q: Even then, couldn't you check if they're the same save file?
A: That's a more complex solution, and a hack. This is a more
maintainable way of solving the problem.