Fixed bug where connect would connect to dead hosts
The `connect` utility was one of the first written for reHack. As such,
it doesn't use any of the helper libraries I've written over time (most
of which are in the data module).

The bug was simple: while it did check whether or not the host was
alive, it didn't DO anything with the result. I've made it so that the
host needs to be alive for connect to accept it.