Added active traces
Active traces are now a thing. No, the game isn't multithreaded, so the
way this works is a bit wonky. But it does work.

Active traces can be three types:

* Corporate (default)
You will be warned and given a day to change your IP address/clear logs.
After a day, the game checks each log in the affected node to see if
your IP address is in it.
* Government
Instant game-over.
* Personal
Nothing yet, but some non-lethal action will occur.

Active traces are broken by clearing logs, which isn't currently
possible, but I will create a `logclear` utility for like $500 that
expunges all logs with your IP address in them.

Active traces can be overriden by another one, causing the previous
trace to be dropped. This is a compromise made for simplicity,
because ignoring new traces would be a MASSIVE exploit waiting
to happen; find a really slow trace and break in all you want.

NOTE: Also added Actions, which can happen on scheduled days. They are
simply functions the game will call later. This is how the Corporate
trace works.