Removed re-licensing clause from CLA
I've come around. I've decided that I don't want to enable a future
scenraio where a bad actor (be it me becoming evil or a corporation
taking over) would be able to abuse the work of others for a profit.

The only reason I added it in the first place was to reduce the
possibility of internal conflict. Upon further consideration,
the risk of abuse is too high for my liking.

I admit it - I made a mistake, and I'm sorry. I will work towards making
sure that I don't make such mistakes again, and I hope to resolve this
issue before it becomes a problem, not once it does.