Added Clone() methods to all VirtualFile/VirtualDirectory/FileSystem
This was a bit of a pain to implement, but it actually works. It means
that every node now has a separate filesystem. Ain't that just neato?

(Also, during development, I had a bit of a panic when my code finally
compiled but got a NullReferenceException on startup. Turns out I'd
removed the code to initialise the root file system, meaning it didn't
exist and would freak out).

Signed-off-by: Gordinator <gordinator@gordinator.org>