Eliminated ALL warnings
With the help of my newly-installed LSP plugins for Neovim, I managed to
track down and squash every single null warning. The one exception was a
warning I had to suppress because a null check made it irrelevant.

In all this excitement, I realised that the User constructor didn't have
its Password argument nullable. As soon as I added the single '?' to
that argument, half the warnings left!

Signed-off-by: Gordinator <gordinator@gordinator.org>