Created news servers and news article loading system
The current system assigns a single .xml file for multiple articles by
multiple servers. This allows for multiple articles to be added to
different news sources simultaneously, which is way better than legacy.

For now, there's only one (1) article so far, and one news server, but
that's to change soon.

Signed-off-by: Gordinator <gordinator@gordinator.org>