Added node health
The BaseNode.CheckHealth() function returns a bool depending on if the
node is healthy. Nodeutils.GetNode() and NodeUtils.GetNodeByAddress()
check this function before returning the node.

Currently, there is no way to "kill" a node, but I'll get around to it
soon enough.

Also added the ReHack Wiki, albeit with no content.

Signed-off-by: Gordinator <gordinator@gordinator.org>