Created financial system
Each node has a balance and a list of transactions performed on it. The
bankutil standard utility lets you send and receive money. A basic
transaction verification system ensures that an account does not send
more money than it has.

The money system will be used as a major focus for the game - missions
have payouts, and that money will be used to purchase goods and
services. And yes, you'll be able to hack into people and steal their
money.

Large organisations will have 'banks' - nodes designed to hold money
securely. They will be juicy targets for hackers, and as such will be
heavily secured.

Signed-off-by: Gordinator <gordinator@gordinator.org>