Overhauled news system
Updated it so that news articles are no longer held in an XML tag - that
was dumb. Instead, that XML tag *references* the article, which is held
in a text file. Much better.

Also, I need to fix the bug that means only one news article appears at
a time. That's annoying.

Signed-off-by: Gordinator <gordinator@gordinator.org>